thermal: mediatek: Add tsensor support for MT67XX/MT68XX thermal system
The auxadc thermal system on MT67XX and MT68XX SoCs seems to sit in
between versions V1 and V2. The downstream driver's version of
raw_to_mcelsius uses a modified formula of the V2 one, but there is a
VTS value for each sensor, slightly adjusted values and slope correction.
Add support for this.

[rva333@protonmail.com: drop efuse parsing since we don't use it]
Signed-off-by: bengris32 <bengris32@protonmail.ch>
Signed-off-by: Roman Vivchar <rva333@protonmail.com>